[Logs Since] Refactor rand_string to helpers/misc.py
